requirements.txt
albumentations==1.2.0
Flask==2.1.2
numpy==1.22.3
pandas==1.4.1
Pillow==9.2.0
torch==1.11.0
